raspberry/python/requirements.txt
pyserial
bleak==0.21.0 
pyserial-asyncio